Move 'IniReader' to 'osu' package
